+++ overlay_container_test.go
@@ -12,12 +12,12 @@ func TestOverlayContainerGet(t *testing.T) {
 		val int
 	}
 
-	container := NewServiceContainer()
+	container := New()
 	container.Set("a", &T{10})
 	container.Set("b", &T{20})
 	container.Set("c", &T{30})
 
-	overlay := Overlay(container, map[interface{}]interface{}{
+	overlay := NewOverlay(container, map[interface{}]interface{}{
 		"a": &T{40},
 		"d": &T{50},
 	})
@@ -50,12 +50,12 @@ func TestOverlayContainerInject(t *testing.T) {
 		D *T1 `service:"d"`
 	}
 
-	container := NewServiceContainer()
+	container := New()
 	container.Set("a", &T1{10})
 	container.Set("b", &T1{20})
 	container.Set("c", &T1{30})
 
-	overlay := Overlay(container, map[interface{}]interface{}{
+	overlay := NewOverlay(container, map[interface{}]interface{}{
 		"a": &T1{40},
 		"d": &T1{50},
 	})