+++ service_container.go
@@ -6,25 +6,6 @@ import (
 	"sync"
 )
 
-// ServiceContainer is a wrapper around services indexed by a unique
-// name. Services can be retrieved by name, or injected into a struct
-// by reading tagged fields.
-type ServiceContainer interface {
-	// Get retrieves the service registered to the given key. It is an
-	// error for a service not to be registered to this key.
-	Get(key interface{}) (interface{}, error)
-
-	// Set registers a service with the given key. It is an error for
-	// a service to already be registered to this key.
-	Set(key interface{}, service interface{}) error
-
-	// Inject will attempt to populate the given type with values from
-	// the service container based on the value's struct tags. An error
-	// may occur if a service has not been registered, a service has a
-	// different type than expected, or struct tags are malformed.
-	Inject(obj interface{}) error
-}
-
 // InjectableServiceKey is an optional interface for service keys.
 //
 // Non-string service key values should implement this interface if they
@@ -36,15 +17,15 @@ type InjectableServiceKey interface {
 	Tag() string
 }
 
-type serviceContainer struct {
+type Container struct {
 	services  map[interface{}]interface{}
 	keysByTag map[string]interface{}
 	mutex     sync.RWMutex
 }
 
-// NewServiceContainer creates an empty service container.
-func NewServiceContainer() ServiceContainer {
-	return &serviceContainer{
+// New creates an empty service container.
+func New() *Container {
+	return &Container{
 		services:  map[interface{}]interface{}{},
 		keysByTag: map[string]interface{}{},
 	}
@@ -52,7 +33,7 @@ func NewServiceContainer() ServiceContainer {
 
 // Get retrieves the service registered to the given key. It is an
 // error for a service not to be registered to this key.
-func (c *serviceContainer) Get(key interface{}) (interface{}, error) {
+func (c *Container) Get(key interface{}) (interface{}, error) {
 	c.mutex.RLock()
 	defer c.mutex.RUnlock()
 
@@ -72,7 +53,7 @@ func (c *serviceContainer) Get(key interface{}) (interface{}, error) {
 
 // Set registers a service with the given key. It is an error for
 // a service to already be registered to this key.
-func (c *serviceContainer) Set(key interface{}, service interface{}) error {
+func (c *Container) Set(key interface{}, service interface{}) error {
 	c.mutex.Lock()
 	defer c.mutex.Unlock()
 
@@ -96,7 +77,7 @@ func (c *serviceContainer) Set(key interface{}, service interface{}) error {
 // the service container based on the value's struct tags. An error
 // may occur if a service has not been registered, a service has a
 // different type than expected, or struct tags are malformed.
-func (c *serviceContainer) Inject(obj interface{}) error {
+func (c *Container) Inject(obj interface{}) error {
 	_, err := inject(c, obj, nil, nil)
 	return err
 }